pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
description = ""
authors = [
    { name = "Gab25", email = "gabrielesorci.25@gmail.com" },
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-google-genai (>=2.1.9,<3.0.0)",
    "langgraph (>=0.6.4,<0.7.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"

